fabs -> fabsf

+++ src/F3D.c
@@ -28,7 +28,7 @@ SOFTWARE.
 #include <math.h>
 #include "ocl_F3D.h"
 
-static const float pi_f = (float) 3.14159265359;
+static const float pi_f = 3.14159265359f;
 
 CVTX_EXPORT bsv_V3f cvtx_F3D_S2S_vel(
 	const cvtx_F3D *self,
@@ -47,7 +47,7 @@ CVTX_EXPORT bsv_V3f cvtx_F3D_S2S_vel(
 	t22 = bsv_V3f_dot(r2, r0) / bsv_V3f_abs(r2);
 	t2 = t21 - t22;
 	/* (NaN != NaN) == TRUE*/
-	return fabs(t1 * t2) <= 3.40282346e38 ? bsv_V3f_mult(crosstmp, t1 * t2) : bsv_V3f_zero();
+	return fabsf(t1 * t2) <= 3.40282346e38 ? bsv_V3f_mult(crosstmp, t1 * t2) : bsv_V3f_zero();
 }
 
 CVTX_EXPORT bsv_V3f cvtx_F3D_S2S_dvort(
